requirements.txt
Django==3.0.6
djangorestframework==3.11.0
django-oauth-toolkit==1.3.2
django-filter==2.2.0
mock==4.0.2
Pillow==7.1.2
django-guardian==2.2.0
celery[redis]==4.4.2
cryptography==2.9.2
twilio==6.40.0
requests==2.23.0
